LLM/LangChain/requirements.txt
openai
langchain
langchain_openai
langchain_experimental
langchainhub
pinecone-client
python-dotenv
tiktoken
docx2txt
pypdf
requests
numpy
pandas
langchain-community
duckduckgo-search
langchain_core
wikipedia
pinecone
langchain_pinecone